Restructure previous reports & remove test reports
