+++ KenticoInspector.Reports.Tests/ApplicationRestartAnalysisTests.cs
@@ -1,71 +1,49 @@
 using KenticoInspector.Core.Models;
 using KenticoInspector.Core.Services.Interfaces;
+using KenticoInspector.Reports.ApplicationRestartAnalysis;
 using KenticoInspector.Reports.Tests.MockHelpers;
 using Moq;
 using NUnit.Framework;
 using System;
 using System.Collections.Generic;
-using System.Dynamic;
 
 namespace KenticoInspector.Reports.Tests
 {
-    [TestFixture(9)]
     [TestFixture(10)]
     [TestFixture(11)]
     [TestFixture(12)]
     public class ApplicationRestartAnalysisTests
     {
-        private Mock<IDatabaseService> _mockDatabaseService;
-
         private Instance _mockInstance;
-
         private InstanceDetails _mockInstanceDetails;
-
+        private Mock<IDatabaseService> _mockDatabaseService;
         private Mock<IInstanceService> _mockInstanceService;
+        private Report _mockReport;
 
         public ApplicationRestartAnalysisTests(int majorVersion)
         {
             InitializeCommonMocks(majorVersion);
+            _mockReport = new Report(_mockDatabaseService.Object, _mockInstanceService.Object);
         }
 
-        [TestCase]
+        [Test]
         public void Should_ReturnEmptyResult_When_DatabaseHasNoEvents()
         {
             // Arrange
             var applicationRestartEvents = new List<ApplicationRestartEvent>();
             _mockDatabaseService
-                .Setup(p => p.ExecuteSqlFromFile<ApplicationRestartEvent>(ApplicationRestartsScripts.GetApplicationRestartEvents))
+                .Setup(p => p.ExecuteSqlFromFile<ApplicationRestartEvent>(Scripts.GetApplicationRestartEvents))
                 .Returns(applicationRestartEvents);
             
-            var report = new ApplicationRestartAnalysis(_mockDatabaseService.Object, _mockInstanceService.Object);
-
             // Act
-            var results = report.GetResults(_mockInstance.Guid);
+            var results = _mockReport.GetResults(_mockInstance.Guid);
 
             // Assert
             Assert.That(results.Data.Rows.Count == 0);
             Assert.That(results.Status == ReportResultsStatus.Information.ToString());
         }
 
-        [TestCase]
-        public void Should_ReturnTableResultType()
-        {
-            // Arrange
-            var applicationRestartEvents = new List<ApplicationRestartEvent>();
-            _mockDatabaseService
-                .Setup(p => p.ExecuteSqlFromFile<ApplicationRestartEvent>(ApplicationRestartsScripts.GetApplicationRestartEvents))
-                .Returns(applicationRestartEvents);
-
-            var report = new ApplicationRestartAnalysis(_mockDatabaseService.Object, _mockInstanceService.Object);
-
-            // Act
-            var results = report.GetResults(_mockInstance.Guid);
-
-            // Assert
-            Assert.That(results.Type == ReportResultsType.Table.ToString());
-        }
-
-        [TestCase]
+        [Test]
         public void Should_ReturnResult_When_DatabaseHasEvents()
         {
             // Arrange
@@ -86,19 +64,32 @@ namespace KenticoInspector.Reports.Tests
             });
 
             _mockDatabaseService
-                .Setup(p => p.ExecuteSqlFromFile<ApplicationRestartEvent>(ApplicationRestartsScripts.GetApplicationRestartEvents))
+                .Setup(p => p.ExecuteSqlFromFile<ApplicationRestartEvent>(Scripts.GetApplicationRestartEvents))
                 .Returns(applicationRestartEvents);
 
-            var report = new ApplicationRestartAnalysis(_mockDatabaseService.Object, _mockInstanceService.Object);
-
             // Act
-            var results = report.GetResults(_mockInstance.Guid);
+            var results = _mockReport.GetResults(_mockInstance.Guid);
 
             // Assert
             Assert.That(results.Data.Rows.Count == 2);
             Assert.That(results.Status == ReportResultsStatus.Information.ToString());
         }
 
+        [Test]
+        public void Should_ReturnTableResultType()
+        {
+            // Arrange
+            var applicationRestartEvents = new List<ApplicationRestartEvent>();
+            _mockDatabaseService
+                .Setup(p => p.ExecuteSqlFromFile<ApplicationRestartEvent>(Scripts.GetApplicationRestartEvents))
+                .Returns(applicationRestartEvents);
+            
+            // Act
+            var results = _mockReport.GetResults(_mockInstance.Guid);
+
+            // Assert
+            Assert.That(results.Type == ReportResultsType.Table.ToString());
+        }
         private void InitializeCommonMocks(int majorVersion)
         {
             _mockInstance = MockInstances.Get(majorVersion);
