+++ KenticoInspector.Reports/ApplicationRestartAnalysis/GetApplicationRestartEvents.sql
