+++ KenticoInspector.Reports/ApplicationRestartAnalysis/Scripts.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace KenticoInspector.Reports.ApplicationRestartAnalysis
+{
+    public static class Scripts
+    {
+        public const string GetApplicationRestartEvents = "ApplicationRestartAnalysis/GetApplicationRestartEvents.sql";
+    }
+}
