+++ KenticoInspector.Reports/ClassTableValidation/ClassWithNoTable.cs
@@ -1,0 +1,9 @@
+﻿namespace KenticoInspector.Reports.ClassTableValidation
+{
+    public class ClassWithNoTable
+    {
+        public string ClassDisplayName { get; set; }
+        public string ClassName { get; set; }
+        public string ClassTableName { get; set; }
+    }
+}
