+++ KenticoInspector.Reports/ClassTableValidation/GetClassesWithNoTable.sql
