+++ KenticoInspector.Reports/ClassTableValidation/GetTablesWithNoClass.sql
