+++ KenticoInspector.Reports/ClassTableValidation/Scripts.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace KenticoInspector.Reports.ClassTableValidation
+{
+    public static class Scripts
+    {
+        public const string ClassesWithNoTable = "ClassTableValidation/GetClassesWithNoTable.sql";
+        public const string TablesWithNoClass = "ClassTableValidation/GetTablesWithNoClass.sql";
+    }
+}
