+++ KenticoInspector.Reports/ClassTableValidation/TableWithNoClass.cs
@@ -1,0 +1,7 @@
+﻿namespace KenticoInspector.Reports.ClassTableValidation
+{
+    public class TableWithNoClass
+    {
+        public string TableName { get; set; }
+    }
+}
